Extract homogeneous numeric expression helpers
